Make EqualsProperty and ColonProperty regexes lazy
Those regexes are used to split property lines into key and value.
But if the value of a property contains an equals sign (=) or a
colon (:) then the regexes fail to split the property lines
correctly. Because they are greedy and try to match as much as
possible. This fix makes them lazy, so they will split at the
first occurrence of (=) or (:).